Disable -Waddress warnings for the unit test targets
GCC 4.2 issues a warning when a string literal's address is used in a conditional statement (and thus will always evaluate to true).
This is triggered by the SenTestCase macros (STAssert*), which check if the macro's description argument is non-null. Due to the code
locality, the compiler determines that the argument may never be false and issues a warning, rather than simply discarding the
intentionally unreachable code path.

These macros are only used in the unit tests; this commit adds -Wno-address to the unit tests' warning flags.


git-svn-id: https://clients.plausible.coop/svn/plcrashreporter/trunk@270 74fe88e9-a195-4aad-b98c-1ad692fed28e