Add internal API for maintaining a linked list of binary images with support for async-safe iteration. Writing may occur concurrently with async-safe reading, but is not async-safe. This will be used to remove the dependency on dyld within the signal handler, which is not defined to be async-safe.
Atomic compare and swap is used to ensure a consistent view of the list for readers. To simplify implementation, a write mutex is held for all updates; the implementation is not designed for efficiency in the face of contention between readers and writers, and it's assumed that no contention should realistically occur.

git-svn-id: https://clients.plausible.coop/svn/plcrashreporter/trunk@350 74fe88e9-a195-4aad-b98c-1ad692fed28e